Created trial fitness test Graph.setFitness(VSet)
Here's an initial implementation of a fitness function.
After messing around with it for a bit, it seemed to me that

[fitness] = [set size]^2 - [connections]^2

... made more sense instead of a linear scaling for connections.

This is because if you only punish connections linearly, you end
up favoring adding connected vertexes for no reason. For example:
- Start with a set of 3 independent vertexes: [fitness] = 9
- Add 1 vertex with 3 connections: [fitness] = 4^2 - 3 = 13

With the [connections]^2 subtracted the second case fitness ends up at
4^2 - 3^2 = 7, which is appropriately lower than the initial fitness

We will probably have to mess around with the fitness as we go, but
this is a start anyways